Move Query Service To V2 Package
